+++ src/frontend/apps/impress/src/features/home/components/HomeHeader.tsx
@@ -17,10 +17,11 @@ export const getHeaderHeight = (isSmallMobile: boolean) =>
 
 export const HomeHeader = () => {
   const { t } = useTranslation();
-  const { themeTokens, spacingsTokens } = useCunninghamTheme();
+  const { themeTokens, spacingsTokens, colorsTokens } = useCunninghamTheme();
   const spacings = spacingsTokens();
   const logo = themeTokens().logo;
   const { isSmallMobile } = useResponsiveStore();
+  const colors = colorsTokens();
 
   return (
     <Box
@@ -44,7 +45,7 @@ export const HomeHeader = () => {
             <ButtonTogglePanel />
           </Box>
         )}
-        {!isSmallMobile && logo && (
+        {!isSmallMobile && logo?.src && (
           <Image
             priority
             src={logo.src}
@@ -61,7 +62,11 @@ export const HomeHeader = () => {
           $position="relative"
           $height="fit-content"
         >
-          <IconDocs aria-label={t('Docs Logo')} width={32} />
+          <IconDocs
+            aria-label={t('Docs Logo')}
+            width={32}
+            color={colors['primary-text']}
+          />
           <Title />
         </Box>
       </Box>